+++ website/app/download/page.tsx
@@ -1,7 +1,7 @@
 import Navbar from "../components/Navbar";
 import { RiMicrosoftFill, RiAppleFill, RiUbuntuLine, RiDownloadLine } from "react-icons/ri";
 
-const VERSION = process.env.NEXT_PUBLIC_VERSION || '1.2.0';
+const VERSION = '1.2.4';
 
 export default function DownloadPage() {
   return (